Implement A-Packages. All classes in duke package, and divided into multiple packages, duke.command, duke.exception and duke.task.
