--- src/main/java/ExitCommand.java
@@ -1,11 +1,0 @@
-public class ExitCommand extends Command {
-
-    public ExitCommand(String cmd) {
-        super(cmd);
-    }
-
-    @Override
-    void execute(Ui ui, TaskList taskList) throws DukeException {
-        ui.showExitMsg();
-    }
-}
