+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,17 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+
+
+public class ExitCommand extends Command {
+
+    public ExitCommand(String cmd) {
+        super(cmd);
+    }
+    @Override
+    public void execute(Ui ui, TaskList taskList) throws DukeException {
+        ui.showExitMsg();
+    }
+}
