+++ src/main/java/duke/command/ListCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Task;
+
+
 public class ListCommand extends Command {
 
     public ListCommand(String cmd) {
@@ -5,9 +13,9 @@ public class ListCommand extends Command {
     }
 
     @Override
-    void execute(Ui ui, TaskList taskList) throws DukeException {
+    public void execute(Ui ui, TaskList taskList) throws DukeException {
         ui.showList();
-        for (int i = 0; i < taskList.getSize(); i ++) {
+        for (int i = 0; i < taskList.getSize(); i++) {
             Task task = taskList.getTask(i);
             if (task == null) {
                 break;
