+++ src/main/java/duke/exception/DukeException.java
@@ -1,6 +1,8 @@
+package duke.exception;
+
 public class DukeException extends Exception {
 
-    public DukeException (String msg) {
+    public DukeException(String msg) {
         super("☹ OOPS!!! " + msg);
     }
 }
