+++ src/main/java/duke/task/Event.java
@@ -1,10 +1,11 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class Event extends Task {
     private LocalDateTime atTime;
-
-    public Event (TaskType type, String name, boolean isMarked, String timeStr) {
+    public Event(TaskType type, String name, boolean isMarked, String timeStr) {
         super(type, name, isMarked);
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HHmm, d/MM/yyyy");
         LocalDateTime atTime = LocalDateTime.parse(timeStr, formatter);
@@ -18,7 +19,6 @@ public class Event extends Task {
     @Override
     public String toString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("hh:mm a, EEE, d MMM yyyy");
-        return "[E]" + super.toString() +
-                " (at: " + atTime.format(formatter) + ")";
+        return "[E]" + super.toString() + " (at: " + atTime.format(formatter) + ")";
     }
 }
