+++ src/main/java/duke/task/Todo.java
@@ -1,6 +1,8 @@
+package duke.task;
+
 public class Todo extends Task {
 
-    public Todo (TaskType type, String name, boolean isMarked) {
+    public Todo(TaskType type, String name, boolean isMarked) {
         super(type, name, isMarked);
     }
